Add soft deletion to products
It's a tradeoff, but this will let us make products appear
as deleted without having to delete purchases or lose the foreign
key reference from purchases to products.
Purchases are a record of activity in the site, and they shouldn't
never be deleted, but they don't make any sense without the sku.
Also we want to retain the capacity of acting as if products
are deleted when they are no longer sold in the platform.